Fix a crash due to simultaneous seat reservation
The game sometimes crashed when two patrons were trying to access a
seat. This should not have happened.
The cause was that two "ReserveSeat" activity were listed at the same
time; this happened because calling "reserving a seat" do not reserve
the seat in itself. You need to have the task ticked once. So, the list
of patrons was iterated, several patrons would try to reserve the same
seat, hence the bug. We need to make sure that tasks are ALWAYS ticked
once at their creation, or we'll get similar bugs.
I added a docstring on the tick method to be sure not to forget about
this one, because it could have been much tougher to find this bug.